Refactor code for easy testing
