+++ src/modules/clearCompleted.js
@@ -1,0 +1,9 @@
+/* eslint-disable import/no-cycle */
+/* eslint-disable import/no-mutable-exports */
+import { tasksArray } from './task.js';
+
+export let filteredArray;
+export default function clearCompleted() {
+  filteredArray = tasksArray.filter((task) => !task.completed);
+  return filteredArray;
+}
